requirements.txt
bidict==0.23.1
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.2
Flask-SocketIO==5.3.6
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
kafka-python==2.0.2
MarkupSafe==2.1.5
pymongo==4.6.3
python-engineio==4.9.0
python-socketio==5.11.2
simple-websocket==1.0.0
Werkzeug==3.0.1
wsproto==1.2.0
